# 开发技术文档/02-API接口定义文档/openapi.yaml
openapi: "3.0.3"
info:
  title: 知微 ziwi SaaS 平台 API
  description: |
    知微(SaaS)平台 — Phase 1 核心业务 API 文档

    ## 认证方式
    - **Bearer JWT**: `Authorization: Bearer <token>` — 用户身份认证
    - **X-Tenant-Id**: 请求头 — 多租户识别（SaaS 模式必传）
    - **X-Deploy-Mode**: 请求头 — 部署模式: saas / onprem / multi-factory
    - **X-Factory-Id**: 请求头 — 工厂ID（multi-factory 模式必传）
    - **X-License-Key**: 请求头 — 许可证密钥（onprem 模式必传）

    ## API 版本
    - URL 路径版本: `/api/v1/{resource}`
    - 同一接口保留 2 个版本，v1 废弃前 3 个月发送 `X-API-Deprecated` 头

    ## 分页规范
    - `page`: 从 1 开始
    - `page_size`: 默认 20，最大 100
    - `total`: 总记录数
    - 响应格式: `{ "items": [], "total": 0, "page": 1, "page_size": 20 }`

    ## 错误响应格式
    ```json
    {
      "code": "400-1001",
      "message": "参数校验失败：必填字段缺失",
      "request_id": "req_abc123"
    }
    ```

    ## 错误码清单
    | code | 说明 |
    |------|------|
    | 400-0000 | 请求参数错误 |
    | 400-1001 | 参数校验失败：必填字段缺失 |
    | 400-1002 | 参数校验失败：格式错误 |
    | 400-1003 | 参数校验失败：值超出允许范围 |
    | 400-2001 | 业务规则冲突：库存不足 |
    | 400-2002 | 业务规则冲突：状态不允许此操作 |
    | 400-2003 | 业务规则冲突：数据已存在（重复） |
    | 401-0000 | 未认证 |
    | 401-1001 | Token 无效或已过期 |
    | 401-1002 | Token 格式错误 |
    | 401-1003 | OAuth 授权失败 |
    | 403-0000 | 无权限 |
    | 403-1001 | 模块未购买（租户未启用该模块） |
    | 403-1002 | 用户角色无此操作权限 |
    | 403-1003 | 数据范围权限不足（不可跨工厂操作） |
    | 404-0000 | 资源不存在 |
    | 404-1001 | 查询的资源 ID 不存在 |
    | 404-1002 | 关联的父资源不存在 |
    | 409-0000 | 数据冲突 |
    | 409-1001 | 并发修改冲突（乐观锁版本号不一致） |
...
    | 429-0000 | 请求过频繁 |
...
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: ["M00-认证"]
      summary: 用户登录
      description: 使用用户名密码或手机验证码登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  example: "P@ssw0rd"
                tenant_id:
                  type: string
                  example: t_12345
              example:
                username: admin
                password: "P@ssw0rd"
                tenant_id: t_12345
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIs...
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
                    example: 86400
                  refresh_token:
                    type: string
                    example: rf_abc123
                  user:
                    $ref: "#/components/schemas/UserInfo"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/refresh:
    post:
      tags: ["M00-认证"]
      summary: 刷新 Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token 刷新成功
  /api/v1/auth/logout:
    post:
      tags: ["M00-认证"]
      summary: 退出登录
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 退出成功
  /api/v1/auth/change-password:
    post:
      tags: ["M00-认证"]
      summary: 修改密码
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: 密码修改成功

  # ==========================================
  # M00 — 租户管理
  # ==========================================
...
  /api/v1/energy/data:
    get:
      tags: ["M11-能碳"]
      summary: 能耗时序数据查询
      operationId: queryEnergyData
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - name: device_id
          in: query
          required: true
          schema:
            type: integer
        - name: metric_name
          in: query
          required: true
          schema:
            type: string
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: aggregation
          in: query
          schema:
            type: string
            enum: [raw, avg, sum, max, min]
            default: raw
        - name: bucket_interval
          in: query
          schema:
            type: string
            example: 1h
      responses:
        "200":
          description: 能耗数据
...
  /api/v1/dashboards/{dashboard_id}/data:
    get:
      tags: ["M13-看板"]
      summary: 看板数据(所有组件数据)
      operationId: getDashboardData
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - name: refresh
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: 看板数据
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token，从登录接口获取"
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: "租户ID（SaaS模式必传）"
    DeployMode:
      type: apiKey
      in: header
      name: X-Deploy-Mode
      description: "部署模式: saas / onprem / multi-factory"
    FactoryId:
      type: apiKey
      in: header
      name: X-Factory-Id
      description: "工厂ID（multi-factory模式必传）"
    LicenseKey:
      type: apiKey
      in: header
      name: X-License-Key
      description: "许可证密钥（onprem模式必传）"
  schemas:
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        real_name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    ApiError:
      type: object
      properties:
        code:
          type: string
          example: "400-1001"
        message:
          type: string
        request_id:
          type: string
...
  responses:
    UnauthorizedError:
      description: 未认证或 Token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            code: "401-1001"
            message: "Token 无效或已过期"
            request_id: "req_abc123"
...
    TooManyRequests:
      description: 请求过频繁
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            code: "429-1001"
            message: "API 调用频率超限（租户级别）"
            request_id: "req_abc123"
...